Phase 1: split storm forcing into met-forcing object model
Introduce Track/StormTrack/ParametricMetForcing/GriddedMetForcing
(surge/{track,parametric,gridded,tools}.py) and port the readers/writers
onto them. Reimplement Storm as a thin backwards-compatibility wrapper: it
holds these objects and exposes every legacy attribute as a delegating
property, preserving the full public surface and the historical import
paths via shims. Demote the stub writers/models off the advertised surface
and split available_models() into Fortran- vs Python-supported registries.

Behavior-neutral by construction: parsing/formatting logic is moved
verbatim, so the Phase 0 gate stays byte-identical with no golden
regeneration (characterization + forcing-aux + bowl-slosh canary +
ike/isaac). Adds tests/test_met_objects.py for the new objects.

Time uses numpy.datetime64 uniformly across track and gridded axes
(met_forcing_refactor.md §6 amended). No Fortran, no surge.data wire, and
no emitted .storm/descriptor bytes change.
